Make Vector/Species/Shape abstract classes and thereby consistent
with the public sub-types for primitive specializations and Mask/Shuffle.
All classes are only extensible from within the same package.
This design decision may be revisted, especially when sealing is possible.
As such the design will not be perturbed to make it hard to swtich back
from classes to interfaces.