8238837: AArch64: TestUpcall failures after JDK-8237358
The binding interpreter changes in JDK-8237358 cause a failure on
AArch64 when a struct parameter whose size is not a multiple of 8 is
spilled to the stack. Because the generated bindings dereference the
struct as a series of Java longs, the final long read will read past the
end of the MemorySegment.

Fix by dereferencing using a type that is appropriate for the size to
copy. E.g. long for 8 bytes, int for 4 bytes, etc.

Tested jdk_foreign on aarch64 and x86_64.

Change-Id: Ic2d5ef11f9306bb14fb789daae7bccd744218bc9