Thanks @Jorn and @Maurizio for your comments.
The templating was cute when I started, but as rows in the table were eliminated it no longer carries its own weight.

This update removes the templating. I decided to keep the implementation in its own class for now, so that we can see if we really want to merge it into MemoryHandles. Most of the implementation is not really sharable with other parts of MemoryHandles.