8247993: AArch64: add support for VaList
Based on the SysV x64 implementation as it's somewhat similar to AArch64.

I added some extra tests to cover some cases on AArch64 that weren't hit
by the existing tests. The new testHugeStructByValue fails on Linux
x86_64 because the HugePoint struct is too large to be passed in
registers in the SysV ABI. Made a minimal fix to pass it on the stack
instead.