8284072: foreign/StdLibTest.java randomly crashes on MacOS/AArch64
This test fails intermittently on Apple silicon macOS with a JVM SIGBUS
and the stack trace below:

  Stack: [0x0000000171758000,0x000000017195b000], sp=0x0000000171957ef0, free space=2047k
  Native frames: (J=compiled Java code, j=interpreted, Vv=VM code, C=native code)
  V [libjvm.dylib+0x984f70] MarkActivationClosure::do_code_blob(CodeBlob*)+0x3c
  V [libjvm.dylib+0x9bfb60] JavaThread::nmethods_do(CodeBlobClosure*)+0xac
  V [libjvm.dylib+0x4669a8] HandshakeState::process_by_self(bool)+0x31c
  V [libjvm.dylib+0x891230] SafepointMechanism::process(JavaThread*, bool)+0xac
  V [libjvm.dylib+0x9f4df0] UpcallLinker::on_entry(UpcallStub::FrameData*)+0xe0
  v blob 0x000000010e1ca4a8
  C [libsystem_c.dylib+0x5c284] _isort+0x88
  C 0x176d80010e0ca444
  j java.lang.invoke.LambdaForm$MH+0x0000000800c82000.invoke(Ljava/lang/Object;JJJJJ)V+16 java.base@19-internal

In UpcallLinker::on_entry we call
ThreadStateTransition::transition_from_native() which needs the code
cache to be writable if it suspends for a safepoint/handshake.

Use the ThreadWXEnable helper to temporarily set the W^X state to W
while inside on_entry() and then toggle it back to X before returning to
the upcall stub.

Tested jdk_foreign plus looped java/foreign/StdLibTest.java for several
hours on an M1 Mac Mini which used to eventually fail.