Add Windows ARM64 ABI support to the Foreign Function & Memory API
There are 2 primary differences between the Windows ARM64 ABI and
the macOS/Linux ARM64 ABI: variadic floating point arguments are passed
in general purpose registers on Windows (instead of the vector registers).
In addition to this, up to 64 bytes of a struct being passed to a
variadic function can be placed in general purpose registers. This
happens regardless of the type of struct (HFA or other generic struct).
This means that a struct can be split across registers and the stack
when invoking a variadic function.

This change introduces tests that compute the sum of the fields of
structs containing 1-4 ints, floats, and doubles to verify that each
field is correctly assigned a register or stack location when invoking
a variadic function (both when the struct can be passed entirely in
registers as well as when the struct spills onto the stack).

For details about the Foreign Function & Memory API, see JEP 434
defined at https://openjdk.org/jeps/434

The Windows ARM64 ABI conventions are documented at
https://learn.microsoft.com/en-us/cpp/build/arm64-windows-abi-conventions